test: reduce app integration context boots
