+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 Upcoming changes focus on CI stability, security hardening, and clearer development guardrails.
 
+- **[dev]** test(performance): **App integration tests reuse fewer Spring contexts.** `BaseIntegrationTest` now owns the shared random-port `@SpringBootTest` and `TestRestTemplate` configuration used by the broad HTMX/app route test group, so those tests no longer declare subtly separate Spring test metadata class by class. The trivial `contextLoads` smoke test now reuses that same base context, and the isolated migration Flyway-configuration equivalence assertion is folded into `MigrationLauncherIntegrationTest` instead of booting as its own class. On the app test task this reduces counted context boots while preserving the existing HTTP and migration coverage.
 - **[dev]** fix(testing): **Shared Postgres shuts down after Spring test contexts.** The test harness now tracks Spring test contexts and closes them before stopping the JVM-wide Testcontainers Postgres container, using a JUnit launcher-session hook with the JVM shutdown hook as a fallback. This removes misleading CI shutdown warnings where Hikari tried to reconnect to a container that had already stopped while beans such as `JobPoller` were still running `@PreDestroy`.
 - **[user]** fix(testing): **Integration-test CI now shuts down shared Postgres cleanly.** The GitHub Actions test run no longer reports misleading Postgres connection-refused errors during shutdown after the real tests have completed.
 - **[dev]** build(security): **Committed secrets are now caught by gitleaks (pre-commit + CI), plus a config-hygiene test.** [gitleaks](https://github.com/gitleaks/gitleaks) is pinned in `.mise.toml` with config/allowlist in `.gitleaks.toml`. A `.husky/pre-commit` hook runs `gitleaks git --staged` (scans only the staged diff, ~0.4s, blocks the commit on a finding); the _Secret Scan_ CI workflow (`.github/workflows/gitleaks.yml`) runs `gitleaks git` over full history on every PR and push to `main` as the hard gate. Complementing it, `NoHardcodedSecretsTest` (in `apps/epistola`'s `app.epistola.suite.architecture`, no Docker, `unitTest`) enforces a project policy gitleaks can't: sensitive `application*.yaml`/`.properties` keys (`password`, `client-secret`, encryption `material`, …) must use `${ENV}`/Secrets rather than a low-entropy literal that entropy-based scanners ignore. A small allowlist of clearly-marked **dev/local** profile files may carry throwaway secrets. See [`docs/testing.md`](docs/testing.md#secret-scanning-gitleaks).
