+++ apps/epistola/src/test/kotlin/app/epistola/suite/BaseIntegrationTest.kt
@@ -6,6 +6,7 @@ import app.epistola.suite.tenants.commands.DeleteTenant
 import app.epistola.suite.tenants.queries.ListTenants
 import app.epistola.suite.testing.IntegrationTestBase
 import org.junit.jupiter.api.BeforeEach
+import org.springframework.boot.resttestclient.autoconfigure.AutoConfigureTestRestTemplate
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.context.annotation.Import
 
@@ -13,14 +14,17 @@ import org.springframework.context.annotation.Import
  * Base class for app-level integration tests.
  *
  * Extends [IntegrationTestBase] with:
- * - Full application context ([EpistolaSuiteApplication]) instead of [TestApplication]
+ * - Full random-port application context ([EpistolaSuiteApplication]) instead of [TestApplication]
+ * - Shared [TestRestTemplate] auto-configuration for HTTP-level app tests
  * - [TestSecurityContextConfiguration] for HTTP request principal binding
  * - Per-class database cleanup via namespaced tenant deletion
  */
 @SpringBootTest(
     classes = [EpistolaSuiteApplication::class],
+    webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT,
     properties = ["epistola.demo.enabled=false"],
 )
+@AutoConfigureTestRestTemplate
 @Import(app.epistola.suite.config.TestSecurityContextConfiguration::class)
 abstract class BaseIntegrationTest : IntegrationTestBase() {
