+++ apps/epistola/src/test/kotlin/app/epistola/suite/EpistolaSuiteApplicationTests.kt
@@ -1,17 +1,8 @@
 package app.epistola.suite
 
-import app.epistola.suite.testing.TestcontainersConfiguration
-import org.junit.jupiter.api.Tag
 import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-import org.springframework.context.annotation.Import
-import org.springframework.test.context.ActiveProfiles
 
-@Import(TestcontainersConfiguration::class)
-@SpringBootTest(classes = [EpistolaSuiteApplication::class])
-@ActiveProfiles("test")
-@Tag("integration")
-class EpistolaSuiteApplicationTests {
+class EpistolaSuiteApplicationTests : BaseIntegrationTest() {
 
     @Test
     fun contextLoads() {
