+++ apps/epistola/src/test/kotlin/app/epistola/suite/catalog/CatalogListHandlerTest.kt
@@ -1,7 +1,6 @@
 package app.epistola.suite.catalog
 
 import app.epistola.suite.BaseIntegrationTest
-import app.epistola.suite.EpistolaSuiteApplication
 import app.epistola.suite.catalog.commands.CreateCatalog
 import app.epistola.suite.catalog.commands.ReleaseCatalogVersion
 import app.epistola.suite.common.ids.CatalogId
@@ -16,8 +15,6 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.resttestclient.TestRestTemplate
-import org.springframework.boot.resttestclient.autoconfigure.AutoConfigureTestRestTemplate
-import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.http.HttpEntity
 import org.springframework.http.HttpHeaders
 import org.springframework.http.HttpMethod
@@ -32,8 +29,6 @@ import org.springframework.util.LinkedMultiValueMap
  * browser fragment parser relocated). Every list mutation must return that
  * wrapper containing *either* the table *or* the empty state.
  */
-@SpringBootTest(classes = [EpistolaSuiteApplication::class], webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@AutoConfigureTestRestTemplate
 class CatalogListHandlerTest : BaseIntegrationTest() {
 
     @Autowired
