+++ apps/epistola/src/test/kotlin/app/epistola/suite/config/HtmxAssetServingIT.kt
@@ -1,13 +1,10 @@
 package app.epistola.suite.config
 
 import app.epistola.suite.BaseIntegrationTest
-import app.epistola.suite.EpistolaSuiteApplication
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.resttestclient.TestRestTemplate
-import org.springframework.boot.resttestclient.autoconfigure.AutoConfigureTestRestTemplate
-import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.boot.web.server.autoconfigure.ServerProperties
 import org.springframework.http.HttpStatus
 
@@ -18,8 +15,6 @@ import org.springframework.http.HttpStatus
  * security chain omits the CSP header; the "no unpkg in rendered HTML" assertion
  * below is the regression guard for self-hosting.
  */
-@SpringBootTest(classes = [EpistolaSuiteApplication::class], webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@AutoConfigureTestRestTemplate
 class HtmxAssetServingIT : BaseIntegrationTest() {
 
     @Autowired
