+++ apps/epistola/src/test/kotlin/app/epistola/suite/handlers/LogsHandlerHtmxTest.kt
@@ -1,7 +1,6 @@
 package app.epistola.suite.handlers
 
 import app.epistola.suite.BaseIntegrationTest
-import app.epistola.suite.EpistolaSuiteApplication
 import app.epistola.suite.common.UUIDv7
 import app.epistola.suite.tenants.Tenant
 import org.assertj.core.api.Assertions.assertThat
@@ -9,8 +8,6 @@ import org.jdbi.v3.core.Jdbi
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.resttestclient.TestRestTemplate
-import org.springframework.boot.resttestclient.autoconfigure.AutoConfigureTestRestTemplate
-import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.http.HttpEntity
 import org.springframework.http.HttpHeaders
 import org.springframework.http.HttpStatus
@@ -22,8 +19,6 @@ import java.time.ZoneOffset
  * Server-contract cover for the tenant Logs viewer: the full page renders, the
  * HTMX search returns the results fragment, and the level filter scopes rows.
  */
-@SpringBootTest(classes = [EpistolaSuiteApplication::class], webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@AutoConfigureTestRestTemplate
 class LogsHandlerHtmxTest : BaseIntegrationTest() {
 
     @Autowired
